Draw box around file and ruler around commit lines
The `diff-so-fancy.rulerWidth` config still applies, but to the ruler
around the commit line.
File lines have a box around them with just the size of the file name
and prefixes/suffixes (e.g. `modified: diff-so-fancy`).
This allows for a better hierarchical view of commits, with commit lines
having a configurable wider ruler around them and file lines ("children"
of commits) having a shorter box around them.